requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
streamlit
Pillow
aspose-ocr-python-net
transformers
torch
scikit-learn
